Major app refactor: Introduced variables.
The variable handling introduced a new behavior: attach of the pattern that holds the context of the elements into each element, and change its state according to this.
This should not happen, but instead values are going to be held in Stretto::Value tokens, and be retrieved at runtime, when needed.
This way, you can create notes individually, but change its properties when attached (for example, the pitch of a note when there is a key signature)